tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.get.double.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.get.double</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.get.double</h1>
<p class="desc">createImageData(w, h) double is converted to long</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("createImageData(w, h) double is converted to long");
_addTest(function(canvas, ctx) {

var imgdata1 = ctx.getImageData(0, 0, 10.01, 10.99);
var imgdata2 = ctx.getImageData(0, 0, -10.01, -10.99);
_assertSame(imgdata1.width, 10, "imgdata1.width", "10");
_assertSame(imgdata1.height, 10, "imgdata1.height", "10");
_assertSame(imgdata2.width, 10, "imgdata2.width", "10");
_assertSame(imgdata2.height, 10, "imgdata2.height", "10");


});
</script>
