LayoutTests/imported/w3c/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.get.double-expected.txt

2d.imageData.get.double
createImageData(w, h) double is converted to long
Actual output:

FAIL createImageData(w, h) double is converted to long assert_equals: imgdata1.width === 10 (got 11[number], expected 10[number]) expected 10 but got 11
